# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management Api Documentation",
        "description": "<b>Api Documentation For</b>, Services responsible for getting the list of all events, get a single event details, creating, updating & deleting an event in database as well as attendee module i.e. getting all attendees for an event, attend an event, get details of an event's attendee & unAttend an Event.",
        "contact": {
            "name": "Aadhar gaur",
            "email": "aadhar41@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://events-management-app.test/api/"
        }
    ],
    "paths": {
        "/events/{event}/attendees": {
            "get": {
                "tags": [
                    "Attendees"
                ],
                "summary": "API (endpoint) for getting all event attendees.",
                "description": "API (endpoint) for getting all event attendees.",
                "operationId": "getEventAttendees",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "relationship data i.e. attendees,user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Error: Not Found. The route events could not be found."
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "format": "object",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "3"
                                                    },
                                                    "user_id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "event_id": {
                                                        "type": "string",
                                                        "example": "10"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-12-16 13:47:06"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-02-11 21:18:46"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/events/10/attendees?include=user,attendees": {
            "post": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Attend an Event",
                "description": "Attend an Event",
                "operationId": "attendEvent",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "relationship data i.e. attendees,user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Id of Event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Error: Not Found. The route events could not be found."
                    },
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "format": "object",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "191"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Aut est minima."
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Ipsam iste hic ut repellat totam est esse ad."
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "2023-12-16 13:47:06"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "2024-02-11 21:18:46"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": {
                                                "id": "3",
                                                "name": "Aut est minima.",
                                                "description": "Ipsam iste hic ut repellat totam est esse ad.",
                                                "start_time": "2023-12-16 13:47:06",
                                                "end_time": "2024-02-11 21:18:46"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{event}/attendees/{attendee}": {
            "get": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Get An Event Attendee Details",
                "description": "Get An Event Attendee Details",
                "operationId": "getEventAttendee",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Id of Event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "attendee",
                        "in": "path",
                        "description": "Id of attendee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Error: Not Found. The route events could not be found."
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "format": "object",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "148"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "America Stanton"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "burnice.cole@example.org"
                                                    },
                                                    "email_verified_at": {
                                                        "type": "string",
                                                        "example": "2023-12-16 13:47:06"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-12-16 13:47:06"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-02-11 21:18:46"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": {
                                                "id": "3",
                                                "user_id": "Aut est minima.",
                                                "event_id": "Ipsam iste hic ut repellat totam est esse ad.",
                                                "created_at": "2023-12-16 13:47:06",
                                                "updated_at": "2024-02-11 21:18:46"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Unattended an Event",
                "description": "Unattended an Event",
                "operationId": "unAttendEvent",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Id of Event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "attendee",
                        "in": "path",
                        "description": "Id of attendee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Error: Not Found. The route events could not be found."
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}